+++ Helpers.h
@@ -1,3 +1,3 @@
 #pragma once
 
-const float SqrtF(float n); // Gives you the square root of a number
+const double SqrtF(float n); // Gives you the square root of a number
